+++ packages/std/src/exports.rs
@@ -20,6 +20,8 @@ use crate::ibc::{
 };
 use crate::imports::{ExternalApi, ExternalQuerier, ExternalStorage};
 use crate::memory::{alloc, consume_region, release_buffer, Region};
+#[cfg(feature = "abort")]
+use crate::panic::install_panic_handler;
 use crate::query::CustomQuery;
 use crate::results::{ContractResult, QueryResponse, Reply, Response};
 use crate::serde::{from_slice, to_vec};
@@ -93,6 +95,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_instantiate(
         instantiate_fn,
         env_ptr as *mut Region,
@@ -121,6 +125,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_execute(
         execute_fn,
         env_ptr as *mut Region,
@@ -148,6 +154,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_migrate(migrate_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32
@@ -170,6 +178,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_sudo(sudo_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32
@@ -191,6 +201,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_reply(reply_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32
@@ -211,6 +223,8 @@ where
     M: DeserializeOwned,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_query(query_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32
@@ -232,6 +246,8 @@ where
     Q: CustomQuery,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_ibc_channel_open(contract_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32
@@ -255,6 +271,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_ibc_channel_connect(contract_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32
@@ -278,6 +296,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_ibc_channel_close(contract_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32
@@ -302,6 +322,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_ibc_packet_receive(contract_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32
@@ -326,6 +348,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_ibc_packet_ack(contract_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32
@@ -351,6 +375,8 @@ where
     C: CustomMsg,
     E: ToString,
 {
+    #[cfg(feature = "abort")]
+    install_panic_handler();
     let res = _do_ibc_packet_timeout(contract_fn, env_ptr as *mut Region, msg_ptr as *mut Region);
     let v = to_vec(&res).unwrap();
     release_buffer(v) as u32